Séparation StaffWheel et ajout stories
