+++ web/src/stubs/WheelEntries.ts
@@ -1,0 +1,41 @@
+export const WheelEntries = {
+  total: 3,
+  documents: [
+    {
+      name: "3 cards",
+      disabled: false,
+      order: 3,
+      probability: 1,
+      $id: "668dafd30002ac1c98da",
+      $createdAt: "2024-07-09T21:46:59.268+00:00",
+      $updatedAt: "2024-07-11T21:55:06.710+00:00",
+      $permissions: [],
+      $databaseId: "6675f377000709b0db07",
+      $collectionId: "668dae8c0038a25a95db",
+    },
+    {
+      name: "1 card",
+      disabled: false,
+      order: 1,
+      probability: 2,
+      $id: "66903cd2003136dded31",
+      $createdAt: "2024-07-11T20:13:06.850+00:00",
+      $updatedAt: "2024-07-11T22:21:16.882+00:00",
+      $permissions: [],
+      $databaseId: "6675f377000709b0db07",
+      $collectionId: "668dae8c0038a25a95db",
+    },
+    {
+      name: "2 cards",
+      disabled: false,
+      order: 2,
+      probability: 2,
+      $id: "66903d040011ca390457",
+      $createdAt: "2024-07-11T20:13:56.325+00:00",
+      $updatedAt: "2024-07-11T21:55:11.639+00:00",
+      $permissions: [],
+      $databaseId: "6675f377000709b0db07",
+      $collectionId: "668dae8c0038a25a95db",
+    },
+  ],
+};
